+++ backend/src/main/java/moheng/recommendtrip/application/RecommendTripService.java
@@ -1,0 +1,7 @@
+package moheng.recommendtrip.application;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class RecommendTripService {
+}
